Move compression stuff out of models.py
